+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - Data is now preloaded at startup so that it is instantly available
 - Adjust the API version of the client so that it works with older versions of Docker
 - Add a way to rename a container
+- Add a search view that allows easy image pull from Docker Hub